Move helper code into lib
