+++ controllers/apply-next/simple/fields.js
@@ -2,8 +2,7 @@
 const { forEach, get, has, isArray, reduce, sumBy, values } = require('lodash');
 const moment = require('moment');
 
-const { Joi, ...commonValidators } = require('../validators');
-const { POSTCODE_PATTERN } = require('../../../modules/postcodes');
+const { Joi, POSTCODE_PATTERN, ...commonValidators } = require('../lib/validators');
 
 const MIN_APPLICANT_AGE = 18;
 const MAX_BUDGET_TOTAL = 10000; // in GBP
